+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Unreleased
+* [#61](https://github.com/fabiodomingues/clj-depend/issues/61): Fix reflection warnings. 
 
 ## 0.11.0 (2024-04-19)
 * [#56](https://github.com/fabiodomingues/clj-depend/issues/56): Added the `:only-ns-in-source-paths` attribute for when it is necessary to consider only namespaces in source paths as part of a layer.